+++ 3days/core/design-system/src/main/res/drawable/ic_pointing_up.png
